move auth features from App directory to Auth module for modularity
